Removed caching lib.
Caching should be done by the user of the library however they please, and as optimized as they wish, rather than a forced one on this library.
This also should result in cleaner code.
Renamed `MySqlDatabaseWrapper` to `DatabaseConnectorWrapper`, as the class might be able to be used in the future for full database abstraction of any connector.